feat: refactor FallbackIgnoreService to use IgnoreListTable and create IgnoreListTable class
